functions/analytics-reporter/requirements.txt
functions-framework>=3.0.0
google-analytics-data>=0.18.0
google-auth>=2.16.0
requests>=2.28.0
